Make CLI command class setup more flexible
